LayoutTests/accessibility/mac/mathml-multiscript-expected.txt

A
1
2

A
1
2
3

A
4
5

A
4
5
6

A
1
2
3
4
5
6

A

A
1
3
5
6
This tests checks that mmultiscript tags will create the correct accessibility math hierarchy.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mscript.role is 'AXRole: AXGroup'
PASS mscript.subrole is 'AXSubrole: AXMathMultiscript'

Check multiscript with one post-script pair
Postscripts:
	0. AXMathSuperscript = AXMathIdentifier
	0. AXMathSubscript = AXMathNumber

Prescripts:


Check multiscript with odd number of post-script pairs
Postscripts:
	0. AXMathSuperscript = AXMathIdentifier
	0. AXMathSubscript = AXMathNumber
	1. AXMathSubscript = AXMathNumber

Prescripts:


Check multiscript with a prescript pair
Postscripts:

Prescripts:
	0. AXMathSuperscript = AXMathIdentifier
	0. AXMathSubscript = AXMathNumber


Check multiscript with odd number of pre-script pairs
Postscripts:

Prescripts:
	0. AXMathSuperscript = AXMathIdentifier
	0. AXMathSubscript = AXMathNumber
	1. AXMathSubscript = AXMathNumber


Check multiscript with pre and post script pairs.
Postscripts:
	0. AXMathSuperscript = AXMathIdentifier
	0. AXMathSubscript = AXMathNumber
	1. AXMathSubscript = AXMathNumber

Prescripts:
	0. AXMathSuperscript = AXMathIdentifier
	0. AXMathSubscript = AXMathNumber
	1. AXMathSubscript = AXMathNumber


Check multiscript with no prescript or postscript pairs.
Postscripts:

Prescripts:


Check multiscript with 'none' tags.
Postscripts:
	0. AXMathSubscript = AXMathNumber
	1. AXMathSubscript = AXMathNumber

Prescripts:
	0. AXMathSuperscript = AXMathIdentifier
	1. AXMathSubscript = AXMathNumber

PASS successfullyParsed is true

TEST COMPLETE
